daemon: split icon retrieval support to its own api_*.go files
