+++ daemon/api_icons_test.go
@@ -1,0 +1,122 @@
+// -*- Mode: Go; indent-tabs-mode: t -*-
+
+/*
+ * Copyright (C) 2014-2020 Canonical Ltd
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+package daemon_test
+
+import (
+	"io/ioutil"
+	"net/http"
+	"net/http/httptest"
+	"os"
+	"path/filepath"
+
+	"gopkg.in/check.v1"
+
+	"github.com/snapcore/snapd/daemon"
+	"github.com/snapcore/snapd/snap"
+)
+
+var _ = check.Suite(&iconsSuite{})
+
+type iconsSuite struct {
+	apiBaseSuite
+}
+
+func (s *iconsSuite) TestAppIconGet(c *check.C) {
+	d := s.daemon(c)
+
+	// have an active foo in the system
+	info := s.mkInstalledInState(c, d, "foo", "bar", "v1", snap.R(10), true, "")
+
+	// have an icon for it in the package itself
+	iconfile := filepath.Join(info.MountDir(), "meta", "gui", "icon.ick")
+	c.Assert(os.MkdirAll(filepath.Dir(iconfile), 0755), check.IsNil)
+	c.Check(ioutil.WriteFile(iconfile, []byte("ick"), 0644), check.IsNil)
+
+	s.vars = map[string]string{"name": "foo"}
+	req, err := http.NewRequest("GET", "/v2/icons/foo/icon", nil)
+	c.Assert(err, check.IsNil)
+
+	rec := httptest.NewRecorder()
+
+	s.req(c, req, nil).ServeHTTP(rec, req)
+	c.Check(rec.Code, check.Equals, 200)
+	c.Check(rec.Body.String(), check.Equals, "ick")
+}
+
+func (s *iconsSuite) TestAppIconGetInactive(c *check.C) {
+	d := s.daemon(c)
+
+	// have an *in*active foo in the system
+	info := s.mkInstalledInState(c, d, "foo", "bar", "v1", snap.R(10), false, "")
+
+	// have an icon for it in the package itself
+	iconfile := filepath.Join(info.MountDir(), "meta", "gui", "icon.ick")
+	c.Assert(os.MkdirAll(filepath.Dir(iconfile), 0755), check.IsNil)
+	c.Check(ioutil.WriteFile(iconfile, []byte("ick"), 0644), check.IsNil)
+
+	s.vars = map[string]string{"name": "foo"}
+	req, err := http.NewRequest("GET", "/v2/icons/foo/icon", nil)
+	c.Assert(err, check.IsNil)
+
+	rec := httptest.NewRecorder()
+
+	s.req(c, req, nil).ServeHTTP(rec, req)
+	c.Check(rec.Code, check.Equals, 200)
+	c.Check(rec.Body.String(), check.Equals, "ick")
+}
+
+func (s *iconsSuite) TestAppIconGetNoIcon(c *check.C) {
+	d := s.daemon(c)
+
+	// have an *in*active foo in the system
+	info := s.mkInstalledInState(c, d, "foo", "bar", "v1", snap.R(10), true, "")
+
+	// NO ICON!
+	err := os.RemoveAll(filepath.Join(info.MountDir(), "meta", "gui", "icon.svg"))
+	c.Assert(err, check.IsNil)
+
+	s.vars = map[string]string{"name": "foo"}
+	req, err := http.NewRequest("GET", "/v2/icons/foo/icon", nil)
+	c.Assert(err, check.IsNil)
+
+	rec := httptest.NewRecorder()
+
+	s.req(c, req, nil).ServeHTTP(rec, req)
+	c.Check(rec.Code/100, check.Equals, 4)
+}
+
+func (s *iconsSuite) TestAppIconGetNoApp(c *check.C) {
+	s.daemon(c)
+
+	s.vars = map[string]string{"name": "foo"}
+	req, err := http.NewRequest("GET", "/v2/icons/foo/icon", nil)
+	c.Assert(err, check.IsNil)
+
+	rec := httptest.NewRecorder()
+
+	s.req(c, req, nil).ServeHTTP(rec, req)
+	c.Check(rec.Code, check.Equals, 404)
+}
+
+func (s *iconsSuite) TestNotInstalledSnapIcon(c *check.C) {
+	info := &snap.Info{SuggestedName: "notInstalledSnap", Media: []snap.MediaInfo{{Type: "icon", URL: "icon.svg"}}}
+	iconfile := daemon.SnapIcon(info)
+	c.Check(iconfile, check.Equals, "")
+}
