+++ daemon/api_test.go
@@ -2104,89 +2104,6 @@ func (s *apiSuite) TestTryChangeConflict(c *check.C) {
 	c.Check(rsp.Result.(*errorResult).Kind, check.Equals, client.ErrorKindSnapChangeConflict)
 }
 
-func (s *apiSuite) TestAppIconGet(c *check.C) {
-	d := s.daemon(c)
-
-	// have an active foo in the system
-	info := s.mkInstalledInState(c, d, "foo", "bar", "v1", snap.R(10), true, "")
-
-	// have an icon for it in the package itself
-	iconfile := filepath.Join(info.MountDir(), "meta", "gui", "icon.ick")
-	c.Assert(os.MkdirAll(filepath.Dir(iconfile), 0755), check.IsNil)
-	c.Check(ioutil.WriteFile(iconfile, []byte("ick"), 0644), check.IsNil)
-
-	s.vars = map[string]string{"name": "foo"}
-	req, err := http.NewRequest("GET", "/v2/icons/foo/icon", nil)
-	c.Assert(err, check.IsNil)
-
-	rec := httptest.NewRecorder()
-
-	appIconCmd.GET(appIconCmd, req, nil).ServeHTTP(rec, req)
-	c.Check(rec.Code, check.Equals, 200)
-	c.Check(rec.Body.String(), check.Equals, "ick")
-}
-
-func (s *apiSuite) TestAppIconGetInactive(c *check.C) {
-	d := s.daemon(c)
-
-	// have an *in*active foo in the system
-	info := s.mkInstalledInState(c, d, "foo", "bar", "v1", snap.R(10), false, "")
-
-	// have an icon for it in the package itself
-	iconfile := filepath.Join(info.MountDir(), "meta", "gui", "icon.ick")
-	c.Assert(os.MkdirAll(filepath.Dir(iconfile), 0755), check.IsNil)
-	c.Check(ioutil.WriteFile(iconfile, []byte("ick"), 0644), check.IsNil)
-
-	s.vars = map[string]string{"name": "foo"}
-	req, err := http.NewRequest("GET", "/v2/icons/foo/icon", nil)
-	c.Assert(err, check.IsNil)
-
-	rec := httptest.NewRecorder()
-
-	appIconCmd.GET(appIconCmd, req, nil).ServeHTTP(rec, req)
-	c.Check(rec.Code, check.Equals, 200)
-	c.Check(rec.Body.String(), check.Equals, "ick")
-}
-
-func (s *apiSuite) TestAppIconGetNoIcon(c *check.C) {
-	d := s.daemon(c)
-
-	// have an *in*active foo in the system
-	info := s.mkInstalledInState(c, d, "foo", "bar", "v1", snap.R(10), true, "")
-
-	// NO ICON!
-	err := os.RemoveAll(filepath.Join(info.MountDir(), "meta", "gui", "icon.svg"))
-	c.Assert(err, check.IsNil)
-
-	s.vars = map[string]string{"name": "foo"}
-	req, err := http.NewRequest("GET", "/v2/icons/foo/icon", nil)
-	c.Assert(err, check.IsNil)
-
-	rec := httptest.NewRecorder()
-
-	appIconCmd.GET(appIconCmd, req, nil).ServeHTTP(rec, req)
-	c.Check(rec.Code/100, check.Equals, 4)
-}
-
-func (s *apiSuite) TestAppIconGetNoApp(c *check.C) {
-	s.daemon(c)
-
-	s.vars = map[string]string{"name": "foo"}
-	req, err := http.NewRequest("GET", "/v2/icons/foo/icon", nil)
-	c.Assert(err, check.IsNil)
-
-	rec := httptest.NewRecorder()
-
-	appIconCmd.GET(appIconCmd, req, nil).ServeHTTP(rec, req)
-	c.Check(rec.Code, check.Equals, 404)
-}
-
-func (s *apiSuite) TestNotInstalledSnapIcon(c *check.C) {
-	info := &snap.Info{SuggestedName: "notInstalledSnap", Media: []snap.MediaInfo{{Type: "icon", URL: "icon.svg"}}}
-	iconfile := snapIcon(info)
-	c.Check(iconfile, check.Equals, "")
-}
-
 func (s *apiSuite) TestInstallOnNonDevModeDistro(c *check.C) {
 	s.testInstall(c, false, snapstate.Flags{}, snap.R(0))
 }
